Created an images folder
We are going to be putting most of our non-auto-generated documentation
into the github wiki feature rather than having it in the repo itself.
The only issue is that images have to be stored in a repo and can't be
uploaded directly to the wiki.

This sets up that folder.